Add ClientsCollection, initial

+++ ClientDB/Trainer.cs
@@ -41,7 +41,7 @@ namespace ClientDB
 				throw new Exception("Error! No such trainer.");
 			}
 
-			m_id = (Int64)long.Parse(userData["id"].ToString());
+			m_id = Int64.Parse(userData["id"].ToString());
 			m_name = userData["name"].ToString();
 			m_phone = userData["phone"].ToString();
 		}
